+++ package.json
@@ -63,7 +63,7 @@
     "eslint-plugin-prettier": "^4.0.0",
     "jest": "^28.1.1",
     "pod-install": "^0.1.0",
-    "prettier": "^2.0.5",
+    "prettier": "^3.4.2",
     "react": "18.2.0",
     "react-native": "0.74.5",
     "react-native-builder-bob": "^0.23.2",
